make N an attribute of the surface class using lazyasd
Ok, this might be a bad idea and there are surely better ways, but this
works for now.

lazyasd is something that scopatz came up with for use in xonsh -- it
lets you define objects but not actually create them until they are
accessed the first time.

We use it in xonsh to improve startup times, mostly with regard to
imports.  We don't import things until we need them and this is made
possible by lazyasd.

In this case, we use N everywhere to represent the number of triangles
in the mesh.  And everywhere in the code we have N = len(surf.triangles)
or something over and over.

But we can't initialize it right away becuase it relies on the value of
surf.triangles which itself is not initialized.

The traditional way to do this (I think) is with an event loop, but I
don't have an event loop sitting around and that seems like overkill.
Instead, I initialize self.N as a LazyObject made up of the length of
self.triangles.

Since self.triangles is assigned before we ever ask for self.N, the
lazyasd lets us get away with this.

This is not super robust, but it works.